// src/app/core/models/chamado.model.ts
export type Chamado = {
  chamadoId: number;
  dataInclusao?: string | null;
  dataUltimaAtualizacao?: string | null;
  titulo?: string | null;
  status?: string | null;
  fantasia?: string | null;
  dataFechamento?: string | null;
  atraso?: number | null;
  reaberto?: string | null;
  usuarioInclusao?: string | null;
  tipo?: string | null;
  dataPrevisao?: string | null;
  responsavel?: string | null;
  qtdeIteracao?: number | null;
  usuarioEmailId?: string | null;
};